Begin migrating AWSMutationCache into promises (#155)
